LayoutTests/imported/w3c/web-platform-tests/css/css-text/white-space/seg-break-transformation-008.tentative-expected.txt


FAIL linebreak only assert_equals: expected 192 but got 199
FAIL spaces linebreak assert_equals: expected 192 but got 199
FAIL linebreak spaces assert_equals: expected 192 but got 199
FAIL spaces linebreak spaces assert_equals: expected 192 but got 199
FAIL multiple linebreaks assert_equals: expected 192 but got 199
FAIL multiple linebreaks + spaces assert_equals: expected 192 but got 199
日本語 ＷＩＤＴＨ
日本語 ＷＩＤＴＨ
日本語 ＷＩＤＴＨ
日本語 ＷＩＤＴＨ
日本語 ＷＩＤＴＨ
日本語 ＷＩＤＴＨ
日本語ＷＩＤＴＨ
